secure_ai_assistant/requirements.txt
# Core dependencies
ollama
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
opencv-python
pillow
pytesseract
shap
lime
watchdog
pydantic
fastapi
uvicorn
sqlite-utils

# Additional dependencies for enhanced functionality
requests
numpy
scikit-image
scikit-learn
matplotlib
pandas
python-dotenv
click
colorama
tqdm